refactor(ledger): collocate the account route registrars :hammer:
Both account registrars move out of account_handler.go into
account_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores.

No route, authz tuple, operation ID or path changed.
